requirements.txt
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
attrs==25.3.0
autopep8==2.3.2
babel==2.17.0
bravado-core==6.1.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
exponent_server_sdk==2.1.0
firebase-admin==6.8.0
fqdn==1.5.1
Genshi==0.7.9
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.72.1
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
isoduration==20.11.0
json2xml==5.0.3
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==5.4.0
msgpack==1.1.0
openpyxl==3.1.5
pep8==1.7.1
pika==1.3.2
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
py3o.formats==0.3
py3o.template==0.10.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.23.0
pyjon.utils==0.7
PyJWT==2.9.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.25.1
rsa==4.9.1
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
swagger-spec-validator==3.0.4
tomli==2.2.1
types-python-dateutil==2.9.0.20250516
typing_extensions==4.13.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.4.0
webcolors==24.11.1
zipp==3.22.0
rest_framework_simplejwt